Correct persistent folder name in Dockerfile
Persistent folder in start_openplc.sh was changed in commit https://github.com/thiagoralves/OpenPLC_v3/commit/a8928f802fe8b39c280ba82ca8805c3adcae2095, but was not changed in Dockerfile.

When docker is running with empty folder for a volume, required files are not created in the volume location.
This commit fixes that.